Reorganize the embed wizard into three steps (#74138)
