+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/components/AuthenticationCard/index.ts
@@ -1,0 +1,1 @@
+export { AuthenticationCard } from "./AuthenticationCard";
