+++ frontend/src/metabase/embedding/embedding-iframe-sdk-setup/components/SdkIframeEmbedSetupModal.tsx
@@ -1,4 +1,3 @@
-import cx from "classnames";
 import { useMemo, useState } from "react";
 import { ResizableBox } from "react-resizable";
 import { match } from "ts-pattern";
@@ -8,8 +7,6 @@ import "react-resizable/css/styles.css";
 
 import noResultsSource from "assets/img/no_results.svg";
 import { useUpdateSettingsMutation } from "metabase/api";
-import CS from "metabase/css/core/index.css";
-import { AuthenticationSection } from "metabase/embedding/embedding-iframe-sdk-setup/components/Authentication/AuthenticationSection";
 import { SdkIframeGuestEmbedStatusBar } from "metabase/embedding/embedding-iframe-sdk-setup/components/SdkIframeGuestEmbedStatusBar";
 import { EMBED_STEPS } from "metabase/embedding/embedding-iframe-sdk-setup/constants";
 import { isQuestionOrDashboardSettings } from "metabase/embedding/embedding-iframe-sdk-setup/utils/is-question-or-dashboard-settings";
@@ -118,18 +115,7 @@ export const SdkIframeEmbedSetupContent = () => {
         <Box className={S.Sidebar} component="aside">
           <Stack className={S.SidebarContent} gap="md">
             <Stack gap="md" flex={1}>
-              <AuthenticationSection />
-
-              <Stack
-                gap="md"
-                flex={1}
-                opacity={allowPreviewAndNavigation ? 1 : 0.5}
-                className={cx(
-                  !allowPreviewAndNavigation && CS.pointerEventsNone,
-                )}
-              >
-                <StepContent />
-              </Stack>
+              <StepContent />
             </Stack>
           </Stack>
